% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metrics.R
\name{class_metrics}
\alias{class_metrics}
\title{classification metrics}
\usage{
class_metrics(ml, train, ml_title)
}
\arguments{
\item{ml}{machine learning algorithm fitted.}

\item{train}{train dataset preprocessed.}

\item{ml_title}{title of the machine learning algorithm.}
}
\value{
A gt object containing the ranking metrics table for the train data and test data.
}
\description{
classification metrics
}
\examples{
1+1
}
